resume_screening_api/requirements-render.txt
Flask==2.3.3
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
nltk==3.8.1
matplotlib==3.8.2
seaborn==0.13.0
Werkzeug==2.3.7
gunicorn==21.2.0
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11
requests==2.31.0
python-dotenv==1.0.0
joblib==1.3.2
lxml==4.9.3
